backend/requirements.txt
PyYAML~=5.1.1
Flask~=2.2.5
